% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{conformal_predict}
\alias{conformal_predict}
\title{Conformal Prediction for Regression}
\usage{
conformal_predict(train_x, train_y, test_x, alpha)
}
\arguments{
\item{train_x}{一个数值型向量，表示训练数据的特征（自变量）}

\item{train_y}{一个数值型向量，表示训练数据的标签（目标值）}

\item{test_x}{一个数值型向量，表示测试数据的特征（自变量）}

\item{alpha}{一个介于 0 和 1 之间的数值，定义预测区间的置信水平}
}
\value{
一个包含每个测试数据点预测区间的列表，包含下界和上界。
}
\description{
本函数基于简单的 conformal prediction 方法为每个测试数据点生成预测区间。
}
\examples{
train_x <- c(1, 2, 3, 4, 5)
train_y <- c(1.1, 2.0, 3.1, 4.2, 5.1)
test_x <- c(2.5, 3.5)
alpha <- 0.05
res <- conformal_predict(train_x, train_y, test_x, alpha)
print(res)
}
